Update README.md
Arrays are used to store multiple different values in the same variable. As long as all of the values are the same type of variable such as all integers or all strings we can store them in an array. This gives us the ability to call any of the values from the array without having to initialize them. Unlike an Array using the Array list class we are able to change the size of the array to fit our needs. When an array is created it has a specific amount of spaces for data, if an array list is used the user can add as many values to the array list as they need. Through my small amount of research it is almost always better to just use an array list when writing code, they are much simpler to manage, although in edge cases the array itself will run a bit smoother if the programmer knows exactly how many objects they are going to use in their code.